add generate recursive pages

+++ main.py
@@ -11,28 +11,39 @@ def main():
     # parser = argparse.ArgumentParser(description="Source Directory to Copy")
     # parser.add_argument("dir", help="The name of the directory to copy in the current directory")
     # args = parser.parse_args()
-    list_files = list_all_files("static")
+
 
     if os.path.exists("public"):
         shutil.rmtree("public")
     os.mkdir("public")
+
+    copy_file("static", "public")
+
+    # generate_page("content/index.md", "template.html", "public/index.html")
+    generate_pages_recursive("content", "template.html", "public")
+
+def copy_file(source: str, dest: str) -> None:
+    if not os.path.exists(source):
+        raise Exception(f"{source} doesn't exist")
+    if not os.path.exists(dest):
+        raise Exception(f"{dest} doesn't exist")
+
+    list_files = list_all_files(source)
+
     for file in list_files:
         copy_file = file.split("/")
         if len(copy_file) > 2:
             file_path = file.split("/")
             req_dir = file_path[1:-1]
-            req_dir = "public/" + "/".join(req_dir)
+            req_dir = f"{dest}/" + "/".join(req_dir)
             os.makedirs(req_dir, exist_ok=True)
             shutil.copy(file, os.path.join(req_dir, copy_file[-1]))
         else:
-            shutil.copy(file, os.path.join("public", copy_file[-1]))
-
-    generate_page("content/index.md", "template.html", "public/index.html")
-
+            shutil.copy(file, os.path.join(dest, copy_file[-1]))
 
 def list_all_files(source: str) -> list[str]:
     if not os.path.exists(source):
-        raise Exception("source dir didn't exist in the current directory")
+        raise Exception(f"{source} didn't exist in the current directory")
 
     res = []
     list_files = os.listdir(source)
@@ -69,6 +80,22 @@ def generate_page(from_path: str, template_path: str, dest_path: str):
     with open(dest_path, "w") as f:
         f.write(full_html)
 
+def generate_pages_recursive(dir_path_content: str, template_path: str, dest_dir_path: str) -> None:
+    print(f"generating page from {dir_path_content} with {template_path} to {dest_dir_path}")
+    if not os.path.exists(dir_path_content):
+        raise Exception(f"source {dir_path_content} doesn't exist")
+    if not os.path.exists(template_path):
+        raise Exception(f"template {template_path} doens't exist")
+
+    if os.path.isdir(dir_path_content):
+        files = os.listdir(dir_path_content)
+        for file in files:
+            new_source = os.path.join(dir_path_content, file)
+            new_dest = os.path.join(dest_dir_path, file)
+            generate_pages_recursive(new_source, template_path, new_dest)
+    else:
+        dest_dir_path = dest_dir_path.replace(".md", ".html")
+        generate_page(dir_path_content, template_path, dest_dir_path)
 
 
 if __name__ == "__main__":
